рефакторинг ( создал вьюмодел и перенёс всю логику туда )
