+++ CareerBoast/Common/Alert/AlertShow.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+struct AlertShow: Identifiable {
+    var id: String { name }
+    let name: String
+}
