--- CareerBoast/Modules/Auth/AuthView.swift
@@ -1,11 +1,0 @@
-import SwiftUI
-
-struct AuthView: View {
-    var body: some View {
-        Text("S")
-    }
-}
-
-#Preview {
-    AuthView()
-}
