--- CareerBoast/Views/SwiftUIView.swift
@@ -1,18 +1,0 @@
-//
-//  SwiftUIView.swift
-//  CareerBoast
-//
-//  Created by Gamıd Khalıdov on 17.03.2024.
-//
-
-import SwiftUI
-
-struct SwiftUIView: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-#Preview {
-    SwiftUIView()
-}
